==> kaggle-data/data-preprocessing/03. csv-clean.ipynb <==
import pandas as pd

df = pd.read_csv('stocklist.csv')

df.index = pd.DatetimeIndex(df['Date'])
df = df.drop(columns=['Date'])

df_weekly = df.resample('W').mean()

df_cleaned = df_weekly.dropna(axis=0, how='all')

df_cleaned.to_csv('cleaned-stocklist.csv')
