==> kaggle-data/data-preprocessing/04. additional_calcs.ipynb <==
import pandas as pd

df = pd.read_csv('cleaned-stocklist.csv')

df = df[1043:].reset_index().drop(columns='index')

df[1:].sum(axis=1)

df[1:].count(axis=1)

df['mkt_avg'] = df.sum(axis=1) / (df.count(axis=1) - 1)

df

diff_df = df.iloc[:, 1:].diff()

diff_df['Date'] = df['Date']

diff_df

normed_vals = [diff_df[i] / diff_df['mkt_avg'] for i in diff_df.columns[:-2]]

normed_df = pd.DataFrame(normed_vals)

normed_df = normed_df.T
normed_df['mkt_avg'] = diff_df['mkt_avg']
normed_df['Date'] = diff_df['Date']

normed_df.columns = diff_df.columns


normed_df.to_csv('stock-differences.csv')
