==> kaggle-data/data-preprocessing/06. additional pruning.ipynb <==
import numpy as np
import pandas as pd
import json
from pathlib import Path
from joblib import parallel_backend

# create an id for each stock:
data_path = Path(r'../data/').glob('*.json')

ticker_dict = {}
j = 0
for i in data_path:
    ticker_dict[i.stem] = j
    j += 1

def index_create(ticker):
    return ticker_dict[ticker]

def null_list(pop_list, j):
    if (data[i.stem]['comparisons'][j]['cor_processed'] == 0):
       pop_list.append(j)
    else:
        pass

with parallel_backend('threading', n_jobs=3):
    for i in data_path:
        try:
            with open(i, 'r') as json_file:
                data = json.load(json_file)
                pop_list = []
                for j in data[i.stem]['comparisons']:
                    null_list(pop_list, j)
                for j in pop_list:
                    data[i.stem]['comparisons'].pop(j)
                json_file.close()
            with open(i, 'w') as json_file:
                json.dump(data, json_file)
        except Exception as e:
            print(e)



def insig_list(pop_list, j):
    if (data[i.stem]['comparisons'][j]['p-val'] >= 0.15):
       pop_list.append(j)
    else:
        pass

with parallel_backend('threading', n_jobs=3):
    for i in data_path:
        try:
            with open(i, 'r') as json_file:
                data = json.load(json_file)
                pop_list = []
                for j in data[i.stem]['comparisons']:
                    insig_list(pop_list, j)
                for j in pop_list:
                    data[i.stem]['comparisons'].pop(j)
                json_file.close()
            with open(i, 'w') as json_file:
                json.dump(data, json_file)
        except Exception as e:
            print(e)

test_path = Path(r'../data-test/').glob('*.json')

def add_reference(j, name):
    data[i.stem]['comparisons'][j]['referencing'] = name

with parallel_backend('threading', n_jobs=3):
    for i in data_path:
        try:
            with open(i, 'r') as json_file:
                data = json.load(json_file)
                for j in data[i.stem]['comparisons']:
                    name = j.split('& ')[1]    
                    add_reference(j, name)
                    try:
                        data[i.stem]['comparisons'][j].pop('cor_processed')
                    except:
                        pass
                json_file.close()
            with open(i, 'w') as json_file:
                json.dump(data, json_file)
        except Exception as e:
            print(e)
