==> kaggle-data/data-preprocessing/08. Rounding.ipynb <==
import numpy as np
from math import isnan
import json
from pathlib import Path

data_path = Path(r'../data-processed').glob('*.json')

test_path = Path(r'../data-test/').glob('*.json')

landing_path = Path(r'../new-data-processed/')

with open(landing_path.joinpath('aieq.json'), 'r') as json_file:
        data = json.load(json_file)
        if isnan(data['aieq']['pos_vals'][0]['p-val']) is False:
            print("Bad")
        else:
            print("Good.")
        
        """
        if "pos_vals" in data[i.stem]:
            for j in range(len(data[i.stem]['pos_vals'])):
                    try:
                        if (data[i.stem]['pos_vals'][j]['p-val'] is not None) & (data[i.stem]['pos_vals'][j]['cor_coef'] is not None):
                            data[i.stem]['pos_vals'][j]['p-val'] = round(data[i.stem]['pos_vals'][j]['p-val'], 6)
                            data[i.stem]['pos_vals'][j]['cor_coef'] = round(data[i.stem]['pos_vals'][j]['cor_coef'], 6)
                        else: 
                            data[i.stem]['pos_vals'].pop(j)
                    except Exception as e:
                        print(e)
        """

for i in data_path:
    with open(i, 'r') as json_file:
        data = json.load(json_file)
        print(i.name)
        for j in data[i.stem]['price']:
            data[i.stem]['price'][j] = round(data[i.stem]['price'][j], 4)

        pos_has_nans = 0
        neg_has_nans = 0
        decor_has_nans = 0

        if "pos_vals" in data[i.stem]:
            for j in range(len(data[i.stem]['pos_vals'])):
                    try:
                        if (isnan(data[i.stem]['pos_vals'][j]['p-val']) is False) & (isnan(data[i.stem]['pos_vals'][j]['cor_coef']) is False):
                            data[i.stem]['pos_vals'][j]['p-val'] = round(data[i.stem]['pos_vals'][j]['p-val'], 6)
                            data[i.stem]['pos_vals'][j]['cor_coef'] = round(data[i.stem]['pos_vals'][j]['cor_coef'], 6)
                        else: 
                            pos_has_nans = 1
                    except Exception as e:
                        print(e)
                          


        if "neg_vals" in data[i.stem]:
            for j in range(len(data[i.stem]['neg_vals'])):
                try:
                    if (isnan(data[i.stem]['neg_vals'][j]['p-val']) is False) & (isnan(data[i.stem]['neg_vals'][j]['cor_coef']) is False):
                        data[i.stem]['neg_vals'][j]['p-val'] = round(data[i.stem]['neg_vals'][j]['p-val'], 6)     
                        data[i.stem]['neg_vals'][j]['cor_coef'] = round(data[i.stem]['neg_vals'][j]['cor_coef'], 6)
                    else:
                        neg_has_nans = 1
                except Exception as e:
                        print(e)
                                     

        if "decorrelated_vals" in data[i.stem]:
            for j in range(len(data[i.stem]['decorrelated_vals'])):

                try:
                    if (isnan(data[i.stem]['decorrelated_vals'][j]['p-val']) is False) & (isnan(data[i.stem]['decorrelated_vals'][j]['cor_coef']) is False):
                        data[i.stem]['decorrelated_vals'][j]['p-val'] = round(data[i.stem]['decorrelated_vals'][j]['p-val'], 6)
                        data[i.stem]['decorrelated_vals'][j]['cor_coef'] = round(data[i.stem]['decorrelated_vals'][j]['cor_coef'], 6)
                    else:
                        decor_has_nans = 1
                except Exception as e:
                        print(e)

        if pos_has_nans == 1:
            data[i.stem].pop('pos_vals')
        if neg_has_nans == 1:
            data[i.stem].pop('neg_vals')
        if decor_has_nans == 1:
            data[i.stem].pop('decorrelated_vals')

        json_file.close()

    with open(landing_path.joinpath(str(i.name)), "w", encoding="utf-8") as new_file:
                        json.dump(data, new_file)    
                        new_file.close()
